Move daceir into the dace folder
All other backends (except the dace backend) have their IR in the
backend-specific subfolder (e.g. `npir.py` inside the `numpy` folder).
Let's do the same for the dace IR.
