+++ src/gt4py/cartesian/gtc/dace/expansion/tasklet_codegen.py
@@ -16,7 +16,7 @@ import dace.subsets
 
 import gt4py.cartesian.gtc.common as common
 from gt4py import eve
-from gt4py.cartesian.gtc import daceir as dcir
+from gt4py.cartesian.gtc.dace import daceir as dcir
 from gt4py.cartesian.gtc.dace.symbol_utils import get_axis_bound_str
 from gt4py.cartesian.gtc.dace.utils import make_dace_subset
 from gt4py.eve.codegen import FormatTemplate as as_fmt
