+++ src/gt4py/cartesian/gtc/dace/expansion_specification.py
@@ -14,7 +14,8 @@ from typing import TYPE_CHECKING, Callable, List, Optional, Set, Union
 
 import dace
 
-from gt4py.cartesian.gtc import common, daceir as dcir, oir
+from gt4py.cartesian.gtc import common, oir
+from gt4py.cartesian.gtc.dace import daceir as dcir
 from gt4py.cartesian.gtc.definitions import Extent
